sql/field.cc:
  Auto merged
sql/mysql_priv.h:
  Auto merged
sql/set_var.cc:
  Auto merged
storage/federated/ha_federated.cc:
  Auto merged
storage/myisammrg/ha_myisammrg.cc:
  Auto merged
plugin/fulltext/plugin_example.c:
  e
  C
sql/field.h:
  e

+++ storage/ndb/src/kernel/blocks/dbdict/printSchemaFile.cpp
@@ -68,7 +68,7 @@ print_head(const char * filename, const SchemaFile * sf)
   if (! checkonly) {
     ndbout << "----- Schemafile: " << filename << " -----" << endl;
     ndbout_c("Magic: %.*s ByteOrder: %.8x NdbVersion: %s FileSize: %d",
-             sizeof(sf->Magic),
+             (int) sizeof(sf->Magic),
              sf->Magic, 
              sf->ByteOrder, 
              version(sf->NdbVersion),
